+++ src/elements/audio_block_format_direct_speakers.cpp
@@ -25,10 +25,13 @@ namespace adm {
       detail::ParameterTraits<SpeakerLabels>::tag) const {
     return speakerLabels_;
   }
-
-  SpeakerPosition AudioBlockFormatDirectSpeakers::get(
-      detail::ParameterTraits<SpeakerPosition>::tag) const {
-    return speakerPosition_.get();
+  CartesianSpeakerPosition AudioBlockFormatDirectSpeakers::get(
+      detail::ParameterTraits<CartesianSpeakerPosition>::tag) const {
+    return boost::get<CartesianSpeakerPosition>(speakerPosition_);
+  }
+  SphericalSpeakerPosition AudioBlockFormatDirectSpeakers::get(
+      detail::ParameterTraits<SphericalSpeakerPosition>::tag) const {
+    return boost::get<SphericalSpeakerPosition>(speakerPosition_);
   }
 
   // ---- Has ---- //
@@ -48,10 +51,13 @@ namespace adm {
       detail::ParameterTraits<SpeakerLabels>::tag) const {
     return speakerLabels_.size() > 0;
   }
-
   bool AudioBlockFormatDirectSpeakers::has(
-      detail::ParameterTraits<SpeakerPosition>::tag) const {
-    return speakerPosition_ != boost::none;
+      detail::ParameterTraits<CartesianSpeakerPosition>::tag) const {
+    return (boost::get<CartesianSpeakerPosition>(&speakerPosition_));
+  }
+  bool AudioBlockFormatDirectSpeakers::has(
+      detail::ParameterTraits<SphericalSpeakerPosition>::tag) const {
+    return (boost::get<SphericalSpeakerPosition>(&speakerPosition_));
   }
 
   // ---- isDefault ---- //
@@ -65,6 +71,14 @@ namespace adm {
   void AudioBlockFormatDirectSpeakers::set(Rtime rtime) { rtime_ = rtime; }
   void AudioBlockFormatDirectSpeakers::set(Duration duration) {
     duration_ = duration;
+  }
+  void AudioBlockFormatDirectSpeakers::set(
+      CartesianSpeakerPosition speakerPosition) {
+    speakerPosition_ = speakerPosition;
+  }
+  void AudioBlockFormatDirectSpeakers::set(
+      SphericalSpeakerPosition speakerPosition) {
+    speakerPosition_ = speakerPosition;
   }
   void AudioBlockFormatDirectSpeakers::set(SpeakerPosition speakerPosition) {
     speakerPosition_ = speakerPosition;
@@ -82,10 +96,6 @@ namespace adm {
   void AudioBlockFormatDirectSpeakers::unset(
       detail::ParameterTraits<SpeakerLabels>::tag) {
     speakerLabels_.clear();
-  }
-  void AudioBlockFormatDirectSpeakers::unset(
-      detail::ParameterTraits<SpeakerPosition>::tag) {
-    speakerPosition_ = boost::none;
   }
 
   // ---- Add ---- //